Properly represent head literal variant
This adds a proper representation of the HeadLiteral variant options
(such that their contents can be accessed) as well as an accessor method
for the variant type. Previously, the head literal type was exposed as
an opaque enum statement, which didn’t provide access to the actual type
of the head literal, including nested elements.

Also, the print_lit method is removed because its functionality is
superceded by the Debug trait implementations.