Fix optional fields
These fields are optional in the Clingo API, that is, they may be set to
null pointers to indicate that these values aren’t set. However, they
were exposed as required fields in these Rust bindings.

This patch turns these fields optional by adjusting the constructor
and accessor methods to accept and return optional values. Some of the
tests are rewritten to account for optional values.

The following fields are turned optional by this patch:

- Aggregate: left_guard and right_guard
- BodyAggregate: left_guard and right_guard
- HeadAggregate: left_guard and right_guard
- TheoryAtom: guard
- TheoryAtomDefinition: guard